feat: MatchingMemberResponse 내부 필드 타입 변경

+++ back-end/src/main/java/com/studymate/backend/matching/dto/MatchingMemberResponse.java
@@ -32,5 +32,5 @@ public class MatchingMemberResponse {
     private int matchingCount;
     private int reviewCount;
     private boolean isLogin;
-    private float matchingPercent;
+    private Double matchingPercent;
 }
